requirements.txt
slack_sdk
PyYAML